+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - False positives in `use_after_free` check when validation errors occur before resource existence checks.
 
+### :wrench: Changed
+
+- Display all cURL commands for stateful failures
+
 ## [4.1.0](https://github.com/schemathesis/schemathesis/compare/v4.0.26...v4.1.0) - 2025-08-17
 
 ### :rocket: Added